fix Valgrind error3

+++ chulkov.leontii/I4/fillingthematrixinspiral.cpp
@@ -10,23 +10,23 @@ int* fillingTheMatrixInSpiral(int* matrix, size_t c, size_t r, size_t indrow, si
   indrow--;
   indcol--;
   matrix[c * indrow + indcol] = 1;
-  size_t val = 1;
-  size_t top = 0, bottom = r - 1, left = 0, right = c - 1;
+  int val = 1;
+  int top = 0, bottom = r - 1, left = 0, right = c - 1;
   while (top <= bottom && left <= right)
   {
-    for (size_t i = top; i <= bottom; i++)
+    for (int i = top; i <= bottom; i++)
     {
       matrix[c * i + left] = val++;
     }
     left++;
-    for (size_t i = left; i <= right; i++)
+    for (int i = left; i <= right; i++)
     {
       matrix[c * bottom + i] = val++;
     }
     bottom--;
     if (left <= right)
     {
-      for (size_t i = bottom; i >= top; i--)
+      for (int i = bottom; i >= top; i--)
       {
         matrix[c * i + right] = val++;
       }
@@ -34,7 +34,7 @@ int* fillingTheMatrixInSpiral(int* matrix, size_t c, size_t r, size_t indrow, si
     }
     if (top <= bottom)
     {
-      for (size_t i = right; i >= left; i--)
+      for (int i = right; i >= left; i--)
       {
         matrix[c * top + i] = val++;
       }
